+++ controllers/users_controller.php
@@ -41,34 +41,52 @@ class usersController{
     }
 
     public function insertUser(){
-        $username = $_REQUEST['userName'];
-        $pass = $_REQUEST['passwd'];
-        $realname = $_REQUEST['realName'];
-        $type = $_REQUEST['userType'];
-
-        $data['insertUser'] = $this->user->addUser($username, $pass, $realname, $type);
-
-        $data['info'] = "Usuario/a añadido/a con éxito.";
-        header("Location:index.php?controller=usersController&action=showUsers&message=" . $data['info']);
+        if($_SESSION['type'] == "admin"){
+            $username = $_REQUEST['userName'];
+            $pass = md5($_REQUEST['passwd']);   //md5() es para cifrar la contraseña en la BD
+            $realname = $_REQUEST['realName'];
+            $type = $_REQUEST['userType'];
+    
+            $data['insertUser'] = $this->user->addUser($username, $pass, $realname, $type);
+    
+            $data['info'] = "Usuario/a añadido/a con éxito.";
+            header("Location:index.php?controller=usersController&action=showUsers&message=" . $data['info']);
+        }
+        else{
+            $data['info'] = "Qué tocas, qué intentas. >:)";
+            header("Location:index.php?controller=resourcesController&action=showResources&message=" . $data['info']);
+        }
     }
 
     public function deleteUser(){
-        $id = $_REQUEST['id'];
-
-        $data['info'] = "Usuario/a eliminado/a con éxito.";
-        $data['delete'] = $this->user->delete($id);
-        header("Location:index.php?controller=usersController&action=showUsers&message=" . $data['info']);
+        if($_SESSION['type'] == "admin"){
+            $id = $_REQUEST['id'];
+    
+            $data['info'] = "Usuario/a eliminado/a con éxito.";
+            $data['delete'] = $this->user->delete($id);
+            header("Location:index.php?controller=usersController&action=showUsers&message=" . $data['info']);
+        }
+        else{
+            $data['info'] = "Qué tocas, qué intentas. >:)";
+            header("Location:index.php?controller=resourcesController&action=showResources&message=" . $data['info']);
+        }
     }
 
     public function changeUser(){
-        if(isset($_SESSION['name'])){
-            $data['name'] = $_SESSION['name'];
+        if($_SESSION['type'] == "admin"){
+            if(isset($_SESSION['name'])){
+                $data['name'] = $_SESSION['name'];
+            }
+            $data['info'] = 'Modificar usuario/a';
+            $data['action'] = 'editUser';
+            $id = $_REQUEST['id'];
+            $data['oldUser'] = $this->user->get($id);
+            View::render("users/add", $data);
         }
-        $data['info'] = 'Modificar usuario/a';
-        $data['action'] = 'editUser';
-        $id = $_REQUEST['id'];
-        $data['oldUser'] = $this->user->get($id);
-        View::render("users/add", $data);
+        else{
+            $data['info'] = "Qué tocas, qué intentas. >:)";
+            header("Location:index.php?controller=resourcesController&action=showResources&message=" . $data['info']);
+        }
     }
 
     public function editUser(){
@@ -89,7 +107,10 @@ class usersController{
             $id = $_SESSION['idUser'];
             $data['name'] = $_SESSION['name'];
             $data['type'] = $_SESSION['type'];
-            $data['info'] = "Bienvenido/a, " . $data['name'];
+            $data['title'] = "Hola, " . $data['name'];
+            if(isset($_REQUEST['message'])){
+                $data['info'] = $_REQUEST['message'];
+            }
             
             $data['userReservations'] = $this->user->getReservations($id);
             View::render("users/user_reservations", $data);
@@ -98,14 +119,23 @@ class usersController{
 
     //función que muestra las reservas de todos los usuarios (para tipo admin)
     public function usersReservations(){
-        if(isset($_SESSION['name']) && isset($_SESSION['type'])){
-            $data['name'] = $_SESSION['name'];
-            $data['type'] = $_SESSION['type'];
+        if($_SESSION['type'] == "admin"){
+            if(isset($_SESSION['name']) && isset($_SESSION['type'])){
+                $data['name'] = $_SESSION['name'];
+                $data['type'] = $_SESSION['type'];
+            }
+            if(isset($_REQUEST['message'])){
+                $data['info'] = $_REQUEST['message'];
+            }
+    
+            $data['allReservations'] = $this->user->getAllReservations();
+            $data['title'] = 'Hola, ' . $data['name'];
+            View::render("users/user_reservations", $data);
         }
-
-        $data['allReservations'] = $this->user->getAllReservations();
-        $data['info'] = 'Bienvenido/a, ' . $data['name'];
-        View::render("users/user_reservations", $data);
+        else{
+            $data['info'] = "Qué tocas, qué intentas. >:)";
+            header("Location:index.php?controller=resourcesController&action=showResources&message=" . $data['info']);
+        }
         
     }
 }
